QPixmap: use 'old' OpenGL-based QPixmap::grabWindow on OS X 10.6.
Commit aaf94776ce1034ec2b2402b9ca0e5cf7c7848b12 introduced use
of the CGDisplayCreateImageForRect API on Mac OS X 10.6 and later
in QPixmap::grabWindow.

This change reverts QPixmap::grabWindow on OS X 10.6 back to its
previous behavior in order to run on VMWare Fusion without crashing.
The old method continues to work fine on 10.6. The new method is only
a necessity on OS X 10.7 and greater, because the old method doesn't
produce a usable image.

While this change allows OS X 10.6 not to use CGDisplayCreateImageForRect(),
there seems to be other problems related to that function call, in particular:

 * A comment on http://qt.gitorious.org/qt/qt/merge_requests/1332
   mentions that the patch crashes on 10.7.2 when using a display
   via Mini Displayport.

 * https://bugreports.qt-project.org/browse/QTBUG-23405
   mentions issues with displays in sleep mode. Note: this
   is unrelated to the issue that this commit is trying to fix.
   The VMWare issue mentioned above doesn't trigger because the
   display is asleep/offline/not alive.